Removing SdkEnvironmentHelper in favor of original SdkEnvironment
