+++ test/unit/managers/PermissionManager.ts
@@ -7,7 +7,7 @@ import Context from '../../../src/models/Context';
 import { NotificationPermission } from '../../../src/models/NotificationPermission';
 import { setUserAgent } from '../../support/tester/browser';
 import Random from '../../support/tester/Random';
-import SdkEnvironmentHelper from '../../../src/helpers/SdkEnvironmentHelper';
+import SdkEnvironment from '../../../src/managers/SdkEnvironment';
 
 test.beforeEach(async t => {
   await TestEnvironment.initialize({
@@ -80,13 +80,13 @@ test('should interpret ambiguous browser permission correctly', async t => {
 test('should detect an insecure top-level frame', async t => {
   const permissionManager = OneSignal.context.permissionManager;
 
-  t.false(SdkEnvironmentHelper.isInsecureOrigin());
+  t.false(SdkEnvironment.isInsecureOrigin());
 
   await TestEnvironment.initialize({
     httpOrHttps: HttpHttpsEnvironment.Http,
     initializeAsIframe: false
   });
-  t.true(SdkEnvironmentHelper.isInsecureOrigin());
+  t.true(SdkEnvironment.isInsecureOrigin());
 });
 
 test('should detect a cross-origin frame-context', async t => {
@@ -120,7 +120,7 @@ test('should not detect an ambiguous permission environment', async t => {
     .stub(permissionManager, 'isCurrentFrameContextCrossOrigin')
     .returns(false);
   const isFrameContextInsecureStub = sinon
-    .stub(SdkEnvironmentHelper, 'isFrameContextInsecure')
+    .stub(SdkEnvironment, 'isFrameContextInsecure')
     .resolves(false);
   t.false(await permissionManager.isPermissionEnvironmentAmbiguous(NotificationPermission.Denied));
   isCurrentFrameContextCrossOriginStub.restore();
@@ -137,7 +137,7 @@ test('should use browser reported permission value in non-ambiguous environment 
     .stub(permissionManager, 'isCurrentFrameContextCrossOrigin')
     .returns(false);
   const isFrameContextInsecureStub = sinon
-    .stub(SdkEnvironmentHelper, 'isFrameContextInsecure')
+    .stub(SdkEnvironment, 'isFrameContextInsecure')
     .resolves(false);
   t.deepEqual(await permissionManager.getNotificationPermission(null), NotificationPermission.Denied);
   isCurrentFrameContextCrossOriginStub.restore();
@@ -153,7 +153,7 @@ test('should detect an ambiguous permission environment', async t => {
     .stub(permissionManager, 'isCurrentFrameContextCrossOrigin')
     .returns(true);
   const isFrameContextInsecureStub = sinon
-    .stub(SdkEnvironmentHelper, 'isFrameContextInsecure')
+    .stub(SdkEnvironment, 'isFrameContextInsecure')
     .resolves(false);
   t.true(await permissionManager.isPermissionEnvironmentAmbiguous(NotificationPermission.Denied));
   isCurrentFrameContextCrossOriginStub.restore();
@@ -165,7 +165,7 @@ test('should detect an ambiguous permission environment', async t => {
       .stub(permissionManager, 'isCurrentFrameContextCrossOrigin')
       .returns(false);
     const isFrameContextInsecureStub = sinon
-      .stub(SdkEnvironmentHelper, 'isFrameContextInsecure')
+      .stub(SdkEnvironment, 'isFrameContextInsecure')
       .resolves(true);
     t.true(await permissionManager.isPermissionEnvironmentAmbiguous(NotificationPermission.Denied));
     isCurrentFrameContextCrossOriginStub.restore();
